backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==41.0.7
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic[email]==2.5.0
jupyter==1.0.0
jupyterlab==4.0.8
notebook==7.0.6
psutil==5.9.6
python-dotenv==1.0.0 